feat: Only rebuild datasets that have changed
This change refactors the command module into Step classes which are
able to determine if they are "dirty", i.e. if they are missing or if
their ingredients have changed since they were last built.

The dirty checks are all based on checksums, rather than timestamps, so
they only care about consistency.

The checksums for data steps include the Python or Jupyter notebook
files containing the code to generate the step, so the step will rebuild
if those files change.

Skipped steps are still listed in the ETL, but they are marked as
"cached".