fix: Include sex code mapping in source checksum
For the garden version of GHE, rename `sex_code_mapping.json` to
`ghe.mapping.sex.json`. The rationale is that any metadata files that a step
uses should use the same stem to support auto detection, even if that
slightly limits reuse.

Updates the `DataStep` class so that all files with the same stem are
included when doing a source checksum.