feat: Make publishing incremental
Instead of doing a single bulk date-based sync to S3, instead we use
checksums in the catalog index to decide dataset by dataset which ones
are out of date.

Datasets that are out of date have a content-based folder sync applied,
where we compare md5 checksums locally against files that are already
there and only upload or delete files as needed to make the folders
sync.