fix: Move repacking to owid-catalog-py
Previously, we attempted to repack data frames into compact formats for
open numbers steps. But it turns out that data types get changed when
they are added to the index. This means it's only worth repacking at the
last minute, when we save the data to disk.

For this reason, repacking was moved into `Table.to_feather()` in
`owid-catalog-py`.