enhance: Speed up github checks for open numbers
Currently when we need to check if a repo is current, we shell out twice
using the `sh` module, once to check the current ref, once to do a
dry-run of `git fetch`. Both of these are slow and mean it takes a long
time to see when the repo is out of date, once it's built.

This change switches to reading from disk the local git sha, and doing
a single HTTPS request using Git's smart HTTPS protocol to Github to get
the latest remote sha.