data(population): Build population from harmonised garden tables
Simplify the recipe for population in `key_indicators` by building it
off pre-harmonized garden tables instead of unharmonized tables.

Note! By default, the Hyde dataset does not provide historical data for
Serbia, only for Serbia and Montenegro together. In order to give Serbia
proper history, we slightly change the Hyde harmonization to map Serbia
and Montenegro to Serbia.

With the above change, the resulting dataset is identical to Ed's
original one.