fix: switch to subprocess to be more ipdb friendly
We were using `ProcessPoolExecutor` to isolate python steps, but it was
not possible to bind a tty and use interactive tools like ipdb for
debugging.

This second approach uses subprocess to shell out to a separate command
for the step, and runs the step within the subprocess. This ensures ipdb
continues to work normally.