fix: massive parquet file for faostat_food_explorer
Fixes #363

A typo in the table metadata for the `faostat_food_explorer` dataset
caused the entire multi-index to be serialised into the metadata,
instead of just the names of the columns.

Fixing the primary key field for this dataset to just store the names of
the columns reduces the parquet file's size from 187M to 61M, much
closer to the feather file.

The feather file was not affected since it follows a different code path.