data: add OECD Regional Wellbeing dataset to meadow
The regional wellbeing dataset is a series of indicators meant to
summarise wellbeing and reported at the subnational level.

This PR loads them in and converts a long form data frame to a wide data
frame (one column per indicator), then stuffs the extra columns into the
metadata for those indicators, simplifying the dataset.

The region codes are special OECD codes, which we would still need to
map to ISO codes if we wanted to be able to standardise them more.