:hammer: Support strict primary key enforcement
This change causes a warning to happen any time you save a table without
a primary key. It also allows a strict mode in which the warning becomes
an exception.

We are trying to move more metadata to the indicator level. The
dimensions an indicator has is a really important piece of metadata that
is make or break for automatically collecting, indexing and reusing the
indicators.

The Buildkite tasks that build production data and do the nightly full
build will not be run in strict mode. However, PRs will be run in strict
mode. This is to avoid any downtime for the ETL, whilst strongly
encouraging future changes to have a primary key.