Update {{cookiecutter.short_name}}.py (#3579)
Setting ds_garden.read(reset_index = False) as default.

Maybe I misunderstand, but I find that the grapher step fails unless I set an index in the grapher step. 

I _think_ it makes sense not to reset the index when reading in the table, but I'm sure I'm missing something!